Split logo into layers
To remove detail for smaller icon versions, we'll want
to get rid of the cloud, and resize the servers and
flames.

Split the logo into layers for each of the three image elements we will
need to manipulate:
* Cloud
* Server & Flames
* Archivist

There is some increased aliasing, especially around the tops of the
flames. However, it shouldn't be noticeable at the icon sizes we're going
to be scaling to.